refactor(Charts-domain-candle-dynamic): 폴더 구조 리팩토링 및 Entity수정
